Add error handling and refactor imports in markdown-remark integration
